refactor: wire domain CLI sub-apps into root app, rename democli → demo
- cli/app.py now imports and registers all domain sub-apps
  (config, credentials, disks, ordering, preflight, snapshots, demo)
  and top-level commands (run, sh)
- cli/__init__.py simplified to just import app + define main()
- democli/ renamed to demo/, pyproject.toml entry point updated

Old cli/ command files remain but are no longer registered.
Tests will be updated in the next commit.
