+++ nbkp/demo/output_cmd.py
