+++ nbkp/demo/seed_cmd.py
